Moves `HTTPClientStub` to a shared scope in a separate file
